+++ one-ll/src/main/scala/onell/algorithms/OnePlusOneEA.scala
@@ -10,7 +10,7 @@ import scala.util.Random
 object OnePlusOneEA extends Algorithm {
   override def name: String = "(1+1)-EA"
   override def metrics: Seq[String] = Seq("Fitness evaluations")
-  override def solve(problem: MutationAwarePseudoBooleanProblem)(implicit rng: Random): Seq[Long] = {
+  override def solve(problem: MutationAwarePseudoBooleanProblem)(implicit rng: Random): Seq[Double] = {
     val n = problem.problemSize
     val m = problem.optimumFitness
     val mutation = new Mutation(n, 1.0 / n)
